app/requirements.txt
gpt-index
flask
flask_cors
python-dotenv